Move story-related image assets to their own subfolder
