+++ assets/images/stories-editor/story-templates/got-logo.png
