+++ assets/images/stories-editor/story-templates/quote-image.png
