+++ assets/src/stories-editor/components/story-controls/index.js
@@ -20,7 +20,7 @@ import { compose } from '@wordpress/compose';
  * Internal dependencies
  */
 import { TemplateInserter } from '../';
-import reorderIcon from '../../../../images/reorder.svg';
+import reorderIcon from '../../../../images/stories-editor/reorder.svg';
 import './edit.css';
 
 function StoryControls( { isReordering, startReordering, saveOrder, resetOrder } ) {
