+++ assets/src/stories-editor/components/template-inserter/index.js
@@ -19,8 +19,8 @@ import { ENTER, SPACE } from '@wordpress/keycodes';
  * Internal dependencies
  */
 import TemplatePreview from './template-preview';
-import pageIcon from '../../../../images/add-page-inserter.svg';
-import addTemplateIcon from '../../../../images/add-template.svg';
+import pageIcon from '../../../../images/stories-editor/add-page-inserter.svg';
+import addTemplateIcon from '../../../../images/stories-editor/add-template.svg';
 import './edit.css';
 import { createSkeletonTemplate, maybeEnqueueFontStyle } from '../../helpers';
